chat-with-docs/pyproject.toml
[project]
name = "chat-with-docs"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "pypdf>=4.0",
  "sentence-transformers>=2.6",
  "faiss-cpu>=1.7.4",
  "numpy>=1.24",
  "httpx>=0.26",
  "unstructured>=0.10.0",
  "python-docx>=0.8.11",
  "python-pptx>=0.6.21",
  "openpyxl>=3.1.0",
  "lxml>=4.9.0",
  "beautifulsoup4>=4.12.0",
  "markdown>=3.4.0",
  "watchdog>=3.0.0",
  "rank-bm25>=0.2.2",
  "openai>=1.0.0",
  "pytest>=8.0.0",
]

[tool.uvicorn]
factory = false
